requirements.txt
pybluez
click
ethereum
pysha3
